backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.62
botocore==1.42.62
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-cors-headers==4.9.0
django-invitations==2.1.0
django-model-utils==5.0.0
django-phonenumber-field==8.4.0
django-tenants==3.10.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.10
langchain-core==1.2.18
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.10
langsmith==0.7.16
lxml==6.0.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
phonenumbers==9.0.24
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyMuPDF==1.27.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
PyYAML==6.0.3
redis==7.2.1
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
vine==5.1.0
wcwidth==0.6.0
xxhash==3.6.0
zstandard==0.25.0
